requirements-backend.txt
fastapi
fastapi-cli
pymongo
pydantic
motor[srv]
beanie
plotly
pandas
gpt4all
pinecone-client
langchain
langchain-community
